requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
astroid==3.3.6
async-timeout==4.0.2
attrs==21.4.0
base58==2.1.1
bitarray==1.2.2
black==24.10.0
borb==2.1.8
bson==0.5.10
camelot-py==0.10.1
cbor==1.0.0
certifi==2022.5.18.1
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.0.12
click==8.1.3
cryptography==39.0.1
cytoolz==0.11.2
dill==0.3.9
distro==1.8.0
et-xmlfile==1.1.0
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
fonttools==4.38.0
frozenlist==1.3.0
ghostscript==0.7
hexbytes==0.2.2
idna==3.3
ipfshttpclient==0.8.0a2
isort==5.13.2
jsonschema==4.5.1
lru-dict==1.1.7
lxml==4.9.2
mccabe==0.7.0
multiaddr==0.0.9
multidict==6.0.2
mypy-extensions==1.0.0
natsort==8.2.0
netaddr==0.8.0
numpy==1.22.4
opencv-python==4.6.0.66
openpyxl==3.1.1
packaging==23.0
pandas==1.5.3
parsimonious==0.8.1
pathspec==0.12.1
pdfminer.six==20221105
Pillow==9.4.0
platformdirs==4.3.6
protobuf==3.20.1
pycparser==2.21
pycryptodome==3.14.1
pylint==3.3.2
PyPDF2==2.9.0
pypng==0.20220715.0
pyrsistent==0.18.1
python-barcode==0.14.0
python-bsonjs==0.3.0
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2022.7.1
qrcode==7.4.2
requests==2.27.1
rlp==2.0.1
scipy==1.9.3
six==1.16.0
solc-select==1.0.3
tabula-py==2.6.0
tabulate==0.9.0
tomli==2.2.1
tomlkit==0.13.2
toolz==0.11.2
typing_extensions==4.5.0
urllib3==1.26.9
varint==1.0.2
web3==5.29.1
websockets==9.1
yarl==1.7.2
